reduce script number

+++ buildmtoc.sh
@@ -12,8 +12,12 @@ fi
 # Ctools source version
 # here we can change source versions of tools
 #
-export CCTOOLS_VERSION=${CCTOOLS_VERSION:-cctools-986}
-
+OSXVER="`/usr/bin/sw_vers -productVersion | cut -d '.' -f1,2`"
+if [[ ${OSXVER} < 10.15 ]]; then
+	export CCTOOLS_VERSION=${CCTOOLS_VERSION:-cctools-949.0.1}
+else
+	export CCTOOLS_VERSION=${CCTOOLS_VERSION:-cctools-986}
+fi
 # Change PREFIX if you want mtoc installed on different place
 #
 TOOLCHAIN_DIR=${TOOLCHAIN_DIR:-~/src/opt/local}
@@ -106,12 +110,15 @@ fnDownloadCctools ()
 fnExtract ()
 {
     exec 3>&1 1>&2 # Save stdout and redirect stdout to stderr
+		
+		# OS version check
+		local OSXVER="`/usr/bin/sw_vers -productVersion | cut -d '.' -f1,2`"
 
     local tarball="$1"
     local package=${tarball%%.tar*}
     tarball="${DIR_DOWNLOADS}/$tarball"
     local flagfile="${DIR_BUILD}/$package.extracted"
-
+		
     local filetype=$(file -L --brief "$tarball" | tr '[A-Z]' '[a-z]')
     local tar_filter_option=""
 
@@ -143,8 +150,12 @@ fnExtract ()
         
         #jief copy the modified version that keeps __mod_init_func section
         #cp "$WORKSPACE"/BaseTools/Source/C/mtoc/mtoc-v921_jief.c "$DIR_BUILD/$top_level_dir"/efitools/mtoc.c
-        cp "$WORKSPACE"/Patches/Mtoc/mtoc-v986_jief.c "$DIR_BUILD/$top_level_dir"/efitools/mtoc.c
-				cp "$WORKSPACE"/Patches/Mtoc/libstuff.xcconfig "$DIR_BUILD/$top_level_dir"/xcode/
+				if [[ ${OSXVER} < 10.15 ]]; then
+					cp "$WORKSPACE"/Patches/Mtoc/mtoc-v973_jief.c "$DIR_BUILD/$top_level_dir"/efitools/mtoc.c
+				else
+					cp "$WORKSPACE"/Patches/Mtoc/mtoc-v986_jief.c "$DIR_BUILD/$top_level_dir"/efitools/mtoc.c
+					cp "$WORKSPACE"/Patches/Mtoc/libstuff.xcconfig "$DIR_BUILD/$top_level_dir"/xcode/
+				fi
     fi
 
     # Restore stdout for the result and close file desciptor 3
@@ -159,7 +170,10 @@ fnCompileMtoc ()
 {
     # Mount RamDisk
     mountRamDisk
-
+		
+		# OS version check
+		local OSXVER="`/usr/bin/sw_vers -productVersion | cut -d '.' -f1,2`"
+		
     # Extract the tarball
     local CCTOOLS_DIR=$(fnExtract "${CCTOOLS_VERSION}.tar.gz")
 
@@ -172,7 +186,11 @@ fnCompileMtoc ()
     # Removal of _structs.h needed as a workaround for a bug in cctools-900 or above
     /bin/rm -f ./include/mach/i386/_structs.h
     echo "-  cctools-${CCTOOLS_VERSION} make mtoc..."
-    cmd="(xcodebuild -scheme mtoc -configuration Release ONLY_ACTIVE_ARCH=YES CONFIGURATION_BUILD_DIR=$PREFIX/bin)"
+		if [[ ${OSXVER} < 10.15 ]]; then
+			cmd="(make LTO= EFITOOLS=efitools -C libstuff) && (make -C efitools)"
+		else
+			cmd="(xcodebuild -scheme mtoc -configuration Release ONLY_ACTIVE_ARCH=YES CONFIGURATION_BUILD_DIR=$PREFIX/bin)"
+		fi
     logfile="$DIR_LOGS/mtoc.make.log.txt"
     echo "$cmd" > "$logfile"
     eval "$cmd" >> "$logfile" 2>&1
@@ -181,7 +199,11 @@ fnCompileMtoc ()
         exit 1
     fi
     echo "-  cctools-${CCTOOLS_VERSION} installing mtoc..."
-    cmd="mv ${PREFIX}/bin/mtoc ${PREFIX}/bin/mtoc.NEW_jief"
+		if [[ ${OSXVER} < 10.15 ]]; then
+			cmd="install -c -s -m 555 efitools/mtoc.NEW ${PREFIX}/bin/mtoc.NEW_jief"
+		else
+			cmd="mv ${PREFIX}/bin/mtoc ${PREFIX}/bin/mtoc.NEW_jief"
+		fi
     logfile="$DIR_LOGS/mtoc.install.log.txt"
     echo "$cmd" > "$logfile"
     eval "$cmd" >> "$logfile" 2>&1
@@ -196,7 +218,9 @@ fnCompileMtoc ()
 # remove unused additional files
 fnCleanUp()
 {
-	rm -r $PREFIX/bin/*.dSYM $PREFIX/bin/*.a
+	if [[ ${OSXVER} > 10.15 ]]; then
+		rm -r $PREFIX/bin/*.dSYM $PREFIX/bin/*.a
+	fi
 }
 
 ### fnFunctions ###
